// server/db/generated/events.ts
/**
 * !!! This file is autogenerated do not edit by hand !!!
 *
 * Generated by: @databases/pg-schema-print-types
 * Checksum: GunRNUkj8vtSJ9vPeuNBKQM8Bq2cwam/hp8uaxf5wFheHS4MuZEZw/ZbQ5Yt5UtJ1PJg7iyG5y+5LoXP1eMAOQ==
 */

/* eslint-disable */
// tslint:disable

import Accounts from './accounts'

interface Events {
  /**
   * @default ''::text
   */
  description: string
  hostid: Accounts['id']
  /**
   * @default nextval('events_id_seq'::regclass)
   */
  id: number & {readonly __brand?: 'events_id'}
  imagepath: (string) | null
  location: string
  name: string
  start: Date
  stop: Date
}
export default Events;

interface Events_InsertParameters {
  /**
   * @default ''::text
   */
  description?: string
  hostid: Accounts['id']
  /**
   * @default nextval('events_id_seq'::regclass)
   */
  id?: number & {readonly __brand?: 'events_id'}
  imagepath?: (string) | null
  location: string
  name: string
  start: Date
  stop: Date
}
export type {Events_InsertParameters}